make browserify a peerDependency
If you install browserify-incremental like this:

    $ npm install browserify-incremental
    $ npm install browserify

You get the following directory hierarchy:

    node_modules/browserify-incremental/node_module/browserify
    node_modules/browserify

With two (possibly divergent, over time) copies of browserify.

But if you do the installation in a different order:

    $ npm install browserify
    $ npm install browserify-incremental

Then your directory hierarchy looks like this:

    node_modules/browserify-incremental
    node_modules/browserify

With just the one copy of browserify.

By specifying browserify as a peerDependency instead, we will always
get the latter directory hierarchy regardless of the order in which
we install the dependencies.